Add text UI

+++ src/BoxUtil/HitBox.cs
@@ -2,5 +2,5 @@ using Godot;
 
 public class HitBox : Area2D
 {
-  [Export] public double Damage = 1.0;
+  [Export] public int Damage = 1;
 }
